correct neutron_subnet's host_route: nexthop should be gateway
The CLI expects the --host-route input option to contain gateway
instead of nexthop. However, when querying the subnet using
--shell format it will still output nexthop.

To make sure that the provider correctly matches input and output we
make sure that it still parses the CLI shell output correctly (nexthop)
but puts it in the dict with the gateway key so it aligns with our
input.